add tests

+++ src/validation_workflow/zip/validation_zip.py
@@ -10,6 +10,8 @@ import os
 
 from system.process import Process
 from system.zip_file import ZipFile
+from system.temporary_directory import TemporaryDirectory
+
 from test_workflow.integ_test.utils import get_password
 from validation_workflow.api_test_cases import ApiTestCases
 from validation_workflow.download_utils import DownloadUtils
@@ -18,8 +20,8 @@ from validation_workflow.validation_args import ValidationArgs
 
 
 class ValidateZip(Validation, DownloadUtils):
-    def __init__(self, args: ValidationArgs) -> None:
-        super().__init__(args)
+    def __init__(self, args: ValidationArgs, tmp_dir: TemporaryDirectory) -> None:
+        super().__init__(args, tmp_dir)
         self.os_process = Process()
         self.osd_process = Process()
 
@@ -51,14 +53,17 @@ class ValidateZip(Validation, DownloadUtils):
                 logging.info(f'All tests Pass : {counter}')
                 return True
             else:
+                self.cleanup()
                 raise Exception(f'Not all tests Pass : {counter}')
         else:
+            self.cleanup()
             raise Exception("Cluster is not ready for API test")
 
     def cleanup(self) -> bool:
         try:
-            self.os_process.terminate()
-            if ("opensearch-dashboards" in self.args.projects):
+            if self.succesful_checks > 0:
+                self.os_process.terminate()
+            if ("opensearch-dashboards" in self.args.projects) and self.succesful_checks == 2:
                 self.osd_process.terminate()
         except:
             raise Exception('Failed to terminate the processes that started OpenSearch and OpenSearch-Dashboards')
